+++ polyphemus/lib/etls/sftp_ingest_cat_to_c4_triage_files_etl.rb
@@ -1,6 +1,9 @@
 require_relative "../db_triage_file_etl"
+require_relative "./ingest_config_module"
 
 class Polyphemus::SftpIngestCatToC4TriageFilesEtl < Polyphemus::DbTriageFileEtl
+  include Polyphemus::WithIngestConfig
+
   def initialize
     @project_name = "c4"
     @bucket_name = "triage"
@@ -38,26 +41,6 @@ class Polyphemus::SftpIngestCatToC4TriageFilesEtl < Polyphemus::DbTriageFileEtl
   end
 
   private
-
-  def cat_config
-    @cat_config ||= sftp_configs.find { |c| c[:alias] == "cat" }
-  end
-
-  def c4_config
-    @c4_config ||= ssh_configs.find { |c| c[:alias] == "c4" }
-  end
-
-  def sftp_configs
-    ingest_configs[:sftp]
-  end
-
-  def ssh_configs
-    ingest_configs[:ssh]
-  end
-
-  def ingest_configs
-    Polyphemus.instance.config(:ingest)
-  end
 
   def update_ingested_timestamp(file_name)
     Polyphemus::IngestFile.where(
